torque: Generate replay in sorted order
Instead of calling the callback directly during the replay, a
generator of the log entries of each type is created. Then,
heapq.merge() is used to iterate over them in order. Since the
heapq_merge() does not support the key argument before Python 3.5,
a wrapper is added that simulated key support.